requirements.txt
# Dependências base — Farollintegrado
# Instale com: pip install -r requirements.txt

# Utilitários
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.1.0

# Supabase
supabase>=2.3.0

# Google APIs
google-api-python-client>=2.114.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0

# OpenAI (opcional — remova se não usar)
openai>=1.12.0
